Re-organise and add tests
