--- openfisca_uk/tests/policy/baseline/finance/tax/income_tax/marriage_allowance.yaml
@@ -1,14 +1,0 @@
-- name: Conditions are met, full Marriage Allowance
-  period: 2021
-  input:
-    people:
-      p1:
-        employment_income: 1_000
-      p2:
-        employment_income: 15_000
-    benunit:
-      adults: [p1, p2]
-      is_married: true
-  output:
-    meets_marriage_allowance_income_conditions: [false, true]
-    marriage_allowance: [0, 1_250]
