+++ openfisca_uk/tests/policy/baseline/hmrc/income_tax/allowances/marriage_allowance.yaml
@@ -1,0 +1,41 @@
+- name: No marriage or civil partner, no Marriage Allowance
+  period: 2021
+  input:
+    people:
+      p1:
+        employment_income: 1_000
+      p2:
+        employment_income: 15_000
+    benunit:
+      adults: [p1, p2]
+      is_married: false
+  output:
+    marriage_allowance: [0, 0]
+
+- name: Conditions are met, full Marriage Allowance
+  period: 2021
+  input:
+    people:
+      p1:
+        employment_income: 1_000
+      p2:
+        employment_income: 15_000
+    benunit:
+      adults: [p1, p2]
+      is_married: true
+  output:
+    marriage_allowance: [0, 1_250]
+
+- name: Earner pays higher rate, no Marriage Allowance
+  period: 2021
+  input:
+    people:
+      p1:
+        employment_income: 1_000
+      p2:
+        employment_income: 51_000
+    benunit:
+      adults: [p1, p2]
+      is_married: true
+  output:
+    marriage_allowance: [0, 0]
