+++ openfisca_uk/variables/finance/tax/income_tax/liability.py
@@ -494,27 +494,6 @@ class is_higher_earner(Variable):
         )
 
 
-class CB_HITC(Variable):
-    value_type = float
-    entity = Person
-    label = "Child Benefit High-Income Tax Charge"
-    definition_period = YEAR
-    reference = "Finance Act 2012 s. 681B"
-
-    def formula(person, period, parameters):
-        CB_received = person.benunit("child_benefit", period, options=[ADD])
-        CB_HITC = parameters(period).tax.income_tax.charges.CB_HITC
-        percentage = (
-            amount_over(
-                person("adjusted_net_income", period),
-                CB_HITC.phase_out_start,
-            )
-            / 1_000
-            * CB_HITC.phase_out_rate
-        )
-        return (percentage * CB_received) * person("is_higher_earner", period)
-
-
 class income_tax(Variable):
     value_type = float
     entity = Person
